requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.12
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.1
